\name{data24.Rd}
\alias{data24.Rd}
\docType{data}
\title{Dataset of characteristics of native corn cobs}
\description{Dataset of characteristics of native Mexican corn cobs obtained from the "Proyecto Global de Maíces Nativos" considering the most relevant variables and a subset of observations for the examples.}
\usage{data("data24")}
\format{
A data frame with 24 observations and 61 variables.
}
\source{
Data available on the website: <https://biodiversidad.gob.mx/diversidad/proyectoMaices>.
}
\references{

CONABIO. 2011. Proyecto Global de Maíces Nativos. Comisión Nacional para el Conocimiento y Uso de la Biodiversidad; Instituto Nacional de Investigaciones Forestales, Agrícolas y Pecuarias; Instituto Nacional de Ecología y cambio Climático. México. https://biodiversidad.gob.mx/diversidad/proyectoMaices.
}
\examples{
data("data24")

}
\keyword{examples,dataset}
