% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data31.R
\docType{data}
\name{data31}
\alias{data31}
\title{Data to show the functioning of imanr functions}
\format{
A data frame with 31 observations and 61 variables.
}
\source{
Data available on the website:
<https://biodiversidad.gob.mx/diversidad/proyectoMaices>.
}
\usage{
data("data31")
}
\description{
Dataset of characteristics of native Mexican corn cobs obtained from the
"Proyecto Global de Maíces Nativos" considering the most relevant variables
and a subset of observations for the examples.
}
\examples{
data(data31)                    # lazy loading
str(data31)

}
\references{
CONABIO. 2011. Proyecto Global de Maíces Nativos.
Comisión Nacional para el Conocimiento y Uso de la Biodiversidad;
Instituto Nacional de Investigaciones Forestales, Agrícolas y Pecuarias;
Instituto Nacional de Ecología y cambio Climático.
México. https://biodiversidad.gob.mx/diversidad/proyectoMaices.
}
\keyword{datasets}
