% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_racialcomplex.R
\name{find_racial_complex}
\alias{find_racial_complex}
\alias{findracialcomplex}
\title{Identify the racial complex of a native corn from Mexico}
\usage{
find_racial_complex(data)
}
\arguments{
\item{data}{A dataset that contains qualitative and quantitative characteristics of a corn or series of corns.
The selected characteristics are related to colors, some measurements and the locations in which
the corn was grown. A template for what has to be filled will be included in the GitHub page of the project.}
}
\value{
\code{findraciacomplex()}
returns a racial complex classification for the dataset or single observation that the user entered
}
\description{
Identify the racial complex of a native corn from Mexico
}
\examples{
findracialcomplex(data20)
}
\seealso{
[impute_data()]
}
\author{
Rafael Nieves-Alvarez (\email{rafles091@gmail.com}), Arturo Sanchez-Porras,
 Aline Romero-Natale, Otilio Arturo Acevedo-Sandoval
}
