% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_data.R
\name{impute_data}
\alias{impute_data}
\title{Impute the data implementing random forest for native corn data.}
\usage{
impute_data(data, useParallel = FALSE)
}
\arguments{
\item{data}{An incomplete dataset that contains qualitative and quantitative
characteristics of a corn or series of corns. The selected characteristics
are related to colors, some measurements, and the locations in which the corn
 was grown. A template for what has to be filled will be included on the
 GitHub page of the project.}

\item{useParallel}{Logical. Perform the analysis in parallel? Defaults to FALSE.}
}
\value{
\code{impute_data()}returns an imputed dataset that can be used with
 \code{find_racialcomplex()}.
}
\description{
Impute and prepare a dataframe to apply the \code{find_racialcomplex()} function,
in case the dataframe has missing values. The imputation is done with random
forests. The database must have the same variables as those in \code{bdMaiz.rds}
in this same package.
}
\examples{
impute_data(data24)
}
\references{
B<c3><a1>ez Vergara, K. J. Estimaci<c3><b3>n de datos faltantes a trav<c3><a9>s de
redes neuronales, una comparaci<c3><b3>n con m<c3><a9>todos simpes y m<c3><ba>ltiples (Doctoral
dissertation, Universidad Santo Tom<c3><a1>s).
}
\seealso{
[find_racialcomplex()]
}
\author{
Rafael Nieves-Alvarez (\email{nievesalvarez1618@gmail.com}), Arturo Sanchez-Porras,
 Aline Romero-Natale, Otilio Arturo Acevedo-Sandoval
}
