% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_racialcomplex.R
\name{find_racial_complex}
\alias{find_racial_complex}
\alias{findracialcomplex}
\title{Identify the Racial Complex of native corn from Mexico}
\usage{
find_racial_complex(data)
}
\arguments{
\item{data}{A dataset that contains qualitative and quantitative characteristics
of a corn ear or series of ears. The selected characteristics are related
 to colors, some measurements and the locations in which the corn was grown.
 A template for what has to be filled is included in the GitHub repo of
 the project.}
}
\value{
\code{findraciacomplex()}
returns a racial complex classification for the dataset or single observation that the user entered
}
\description{
Identify the Racial Complex of native corn from Mexico
}
\examples{
\dontrun{
find_racial_complex(data24)#'
}

}
\references{
James, G., Witten, D., Hastie, T., & Tibshirani, R. (2013).
An introduction to statistical learning: With applications in R (1st ed.). Springer.\\
Kuhn, M., & Johnson, K. (2013). Applied predictive modeling (1st ed.). Springer.\\
Monroy, L. G. D. (2007). Estadística Multivariada: Inferencia y Métodos. Univ. Nacional.
}
\seealso{
[impute_data()]
}
\author{
Rafael Nieves-Alvarez (\email{nievesalvarez1618@gmail.com}), Arturo Sanchez-Porras,
 Aline Romero-Natale, Otilio Arturo Acevedo-Sandoval
}
